+++ src/Resources/contao/config/config.php
@@ -1,3 +1,0 @@
-<?php
-
-$GLOBALS['TL_HOOKS']['generatePage'][]	= [\HeimrichHannot\BootstrapHoverDropdownBundle\EventListener\GeneratePageListener::class, 'onGeneratePage'];